- switch to cockroach db - remove all last xodus related stuff

+++ src/main/kotlin/com/jalgoarena/data/UserRepository.kt
@@ -1,9 +1,10 @@
 package com.jalgoarena.data
 
+import com.jalgoarena.domain.User
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.*
 
-interface UserRepository : JpaRepository<User, Long> {
+interface UserRepository : JpaRepository<User, Int> {
     fun findByUsername(username: String): Optional<User>
     fun findByEmail(email: String): Optional<User>
 }
